backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
neo4j==5.16.0
openai>=1.12.0
umap-learn>=0.5.5
scikit-learn>=1.6.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
numpy>=1.26.0,<2.0.0
